% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_power.R
\name{bcp}
\alias{bcp}
\title{Weighted conditional power over all clinically relevant deltas}
\usage{
bcp(z1, c2, n2, cf, ce, delta.mcr)
}
\arguments{
\item{z1}{the observed value of the first stage test statistic}

\item{c2}{a vector with c2-values}

\item{n2}{a vector with n2-values}

\item{cf}{stopping for futility boundary}

\item{ce}{stopping for efficacy boundary}

\item{delta.mcr}{minimal clinically relevant effect}
}
\description{
Compute the conditional power for a weighted average of
all delta-values which are clinically relevant
}
