% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_power.R
\name{cp}
\alias{cp}
\title{Conditional power for specific delta}
\usage{
cp(z1, c2, n2, cf, ce, delta)
}
\arguments{
\item{z1}{the observed value of the first stage test statistic}

\item{c2}{a vector with c2-values}

\item{n2}{a vector with n2-values}

\item{cf}{stopping for futility boundary}

\item{ce}{stopping for efficacy boundary}

\item{delta}{effect size}
}
\description{
Compute the conditional power for a specified effect size
and speficied parameters
}
