% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{dn2.l1}
\alias{dn2.l1}
\title{L_1 - norm of the derivative of n2}
\usage{
dn2.l1(cf, ce, n2)
}
\arguments{
\item{cf}{Stopping for futility boundary}

\item{ce}{Stopping for efficacy boundary}

\item{n2}{Vector of n2-values}
}
\description{
Approximates the L_1-norm of the derivative of n2 w.r.t. z1.
}
