% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_power.R
\name{ecp}
\alias{ecp}
\title{Expected conditional power}
\usage{
ecp(n1, cf, ce, n2, c2, delta.mcr, weighted.alternative, delta.alt)
}
\arguments{
\item{n1}{first stage sample size}

\item{cf}{stopping for futility boundary}

\item{ce}{stopping for efficacy boundary}

\item{n2}{vector with n2-values}

\item{c2}{vector with c2-values}

\item{delta.mcr}{minimal clinically relevant effect size}

\item{weighted_alternative}{Should a weighted alternative be used?}
}
\description{
Compute the expected conditional power weighted over all
minimal clinically relevant effect sizes or for a
point alternative
}
