% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{ess}
\alias{ess}
\title{Expected sample size}
\usage{
ess(n1, cf, ce, n2, weighted.alternative, delta.alt)
}
\arguments{
\item{n1}{First stage sample size}

\item{cf}{stoppinf for futility boundary}

\item{ce}{Stopping for efficacy boundary}

\item{n2}{Vector with n2-values}

\item{weighted_alternative}{Should a weighted average be used?}
}
\description{
Computes the expected sample size under a prespecified (average of)
effect sizes
}
