% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_design}
\alias{opt_design}
\title{Compute an optimal design}
\usage{
opt_design(alpha, beta, lambda, weighted.alternative = FALSE,
  delta.mcr = 0, delta.alt = 0.3, tau = 0.1, n.max = Inf)
}
\arguments{
\item{alpha}{Maximal type one error rate}

\item{beta}{Maximal type two error rate}

\item{lambda}{Two-dimensional vector with penalties for CP and ||n_2^'||_1}

\item{weighted.alternative}{Should a weighted alternative be used?}

\item{delta.mcr}{Minimal clinically relevant effect size}

\item{delta.alt}{Point alternative effect size if weighted.alternative = F,
prior mean otherwise}

\item{tau}{Standard deviation of the prior density}

\item{n.max}{Maximal sample size per stage

Smoothing splines are used to approximate n_2 and c_2 functions.
All integrals are approximates by Boole's Newton-Cotes formula
or by the \code{R}-routine \code{integrate}.}
}
\description{
\code{opt_design} returns a design that is optimal for
specified parameters and a desired mix of expected sample size,
its variability and conditional power and its variability
}
