% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot important key figures of a two-stage design}
\usage{
plot_design(d)
}
\arguments{
\item{d}{An object of class design}
}
\description{
Returns 6 plots for the first-stage,
n_2-function, c_2-function,
power-curve, conditional power curve
and expected sample size
}
