% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{pow}
\alias{pow}
\title{Power}
\usage{
pow(n1, cf, ce, n2, c2, delta.mcr, weighted.alternative, delta.alt, tau)
}
\arguments{
\item{n1}{First stage sample size}

\item{cf}{stoppinf for futility boundary}

\item{ce}{Stopping for efficacy boundary}

\item{n2}{Vector with n2-values}

\item{c2}{Vector with c2-values}

\item{delta.mcr}{Minimal clinically relevant effect size}

\item{weighted_alternative}{Should a weighted alternative be used?}
}
\description{
Compute the power of a design
}
