% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{score}
\alias{score}
\title{Compute the score}
\usage{
score(n1, cf, ce, n2, c2, lambda, weighted.alternative, delta.mcr,
  delta.alt, tau)
}
\arguments{
\item{n1}{First-stage sample size}

\item{cf}{Stopping for futility boundary}

\item{ce}{Stopping for efficacy boundary}

\item{n2}{Vector with n2-values}

\item{c2}{Vector with c2-values}

\item{lambda}{Two-dimensional vector with penalties for CP and ||n_2^'||_1}

\item{weighted.alternative}{Should a weighted alternative be used?}

\item{delta.mcr}{Minimal clinically relevant effect size}

\item{delta.alt}{Point alternative effect size if weighted.alternative = F,
prior mean otherwise}

\item{tau}{Standard deviation of prior density}
}
\description{
score returns the weighted score
}
