% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{toe}
\alias{toe}
\title{Type one error rate}
\usage{
toe(cf, ce, c2, n1)
}
\arguments{
\item{cf}{stopping for futility boundary}

\item{ce}{stopping for efficacy boundary}

\item{c2}{vector with c2-values}
}
\description{
Compute the type one error for given decision boundaries
}
